app/requirements.txt
psycopg2-binary
tensorflow==2.13.0
pandas
numpy
boto3
ydf